# api/swagger.yaml
openapi: 3.0.3
info:
  title: shiftbase API
  description: Shift scheduling & attendance REST API (data contoh fiktif).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: lokal (docker compose)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security: []
paths:
...
  /v1/reports/overtime:
    get:
      summary: Lembur per employee (admin, manager)
      operationId: overtime
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: "[{employee_id,name,total_hours,overtime_hours}]" }
  /v1/reports/coverage:
    get:
      summary: Headcount per tanggal (admin, manager)
      operationId: coverage
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: "[{date,headcount}]" }